api/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pika>=1.3.0
httpx>=0.28
pydantic>=2.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
python-multipart>=0.0.9
python-jose>=3.0.0
